requirements.txt
Django
djangorestframework
langchain
openai
faiss-cpu
tiktoken
langchain-community
langchain-openai



